add exception handler

+++ src/main/java/com/main/library/service/impl/BookServiceImpl.java
@@ -7,17 +7,22 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.main.library.entity.Book;
+import com.main.library.entity.BorrowingRecord;
+import com.main.library.exception.CustomException;
 import com.main.library.repo.BookRepo;
 import com.main.library.service.BookService;
+import com.main.library.service.BorrowingRecordService;
 
 @Service
-public class BookServiceImpl implements BookService{
+public class BookServiceImpl implements BookService {
 
 	@Autowired
 	private BookRepo bookRepo;
-	
+	@Autowired
+	private BorrowingRecordService recordService;
+
 	@Override
-	public Optional<Book> findById(Long id) {		
+	public Optional<Book> findById(Long id) {
 		return bookRepo.findById(id);
 	}
 
@@ -28,17 +33,29 @@ public class BookServiceImpl implements BookService{
 
 	@Override
 	public void deleteById(Long id) {
-		bookRepo.deleteById(id);	
+		List<BorrowingRecord> records = recordService.findByBookId(id);
+		if (records.size() > 0 && records.get(0).getReturnDate() == null) {
+			throw new CustomException("this Book already Borrowed to " + records.get(0).getPatron().getName());
+		}
+		if (!records.isEmpty()) {
+			recordService.deleteAll(records);
+		}
+		bookRepo.deleteById(id);
 	}
 
 	@Override
-	public Book updateBook(Book book) {	
+	public Book updateBook(Book book) {
 		return bookRepo.save(book);
 	}
 
 	@Override
 	public Book save(Book book) {
 		return bookRepo.save(book);
+	}
+
+	@Override
+	public Book findByIsbn(String isbn) {
+		return bookRepo.findByIsbn( isbn);
 	}
 
 }
